refactor: setting 디렉터리로 설정 관련 스크린 이동
